refactor: extract session renewal out of main.ts
Last extraction in the main.ts split. 3,151 lines to 2,618, and 6,143 at the
start of the series.

session-renewal.ts owns the two timers that keep a Plaud session alive
without the user thinking about it: the pre-expiry warning (#78) and the
silent refresh, plus the notices shown when they cannot. Six pieces of state
move with them (two timer ids, two notice handles, the in-flight and failed
flags). They used to sit among the plugin's other twenty-odd fields where
nothing stopped an unrelated method reaching into them; they are now one
object with one teardown, which is what the ordering discipline in these
methods always assumed.

The host interface is long, at fifteen members, and that is the honest
shape: renewal reads the credential, writes settings, posts notices, defers
to an open sign-in, and hands a freshly minted token to the capture store.
Naming them is the point. Before this, every one was an unannounced reach
into the plugin and the ordering rules depended on all of them silently.

Four members stay on the plugin as forwarders, because they are its
published surface: reconcileSessionExpiryWarning and reconcileSessionRefresh
(the capture store reconciles both after every credential write),
canRenewCredential and refreshSessionNow (the settings tab reaches the first
through SettingsTabHost, the debug command the second). Forwarding rather
than handing callers the renewal object keeps its state private to it.

onunload's two inline timer clears become sessionRenewal.dispose(), which
does the same two clears plus the notice handles.

The moved code is TOKEN-IDENTICAL to the original, verified by normalizing
the host rewrite away and comparing 22,773 characters. Prettier re-wrapped
it at the new nesting depth; nothing else changed.

No behavior change. 1,211 tests pass, unchanged.
